+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### CHANGELOG
 
+#### Version 0.3.0
+
+* Added a border around the tasks
+* Moved the "bottom line" code to ActiveLine.qml, and added a progress bar effect to it.
+
 #### Version 0.2.0
 
 * Fixed the appearance of "more than one window effect"